requirements.txt
scikit-learn==1.6.0
kafka-python==2.0.2
beautifulsoup4==4.12.3
requests==2.32.3